% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{MERIDIAN application server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal Shiny objects.}
}
\description{
MERIDIAN application server
}
\keyword{internal}
