% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{build_ge_matrices}
\alias{build_ge_matrices}
\title{Build fast GxE mean matrix and additive interaction matrix}
\usage{
build_ge_matrices(df_std, gen_col, env_col, trait)
}
\arguments{
\item{df_std}{Cleaned data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{trait}{Trait column name}
}
\value{
List with ge_means, interaction, means, dims and labels
}
\description{
Build fast GxE mean matrix and additive interaction matrix
}
