% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{compute_ammi_metrics_fast}
\alias{compute_ammi_metrics_fast}
\title{Fast AMMI metrics (ASV, WAAS, SIPC) from interaction matrix}
\usage{
compute_ammi_metrics_fast(ge)
}
\arguments{
\item{ge}{GxE matrix object from build_ge_matrices()}
}
\value{
Data frame of AMMI stability metrics
}
\description{
Fast AMMI metrics (ASV, WAAS, SIPC) from interaction matrix
}
