% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{compute_blues}
\alias{compute_blues}
\title{Compute BLUEs (genotype as fixed, environment as random)}
\usage{
compute_blues(df, gen_col, env_col, rep_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{rep_col}{Replicate column name}

\item{trait}{Trait column name}
}
\value{
Data frame with Genotype, BLUE, SE, CI_lower, CI_upper
}
\description{
Compute BLUEs (genotype as fixed, environment as random)
}
