% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{compute_blups}
\alias{compute_blups}
\title{Compute BLUPs (genotype as random, environment as fixed)}
\usage{
compute_blups(df, gen_col, env_col, rep_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{rep_col}{Replicate column name}

\item{trait}{Trait column name}
}
\value{
Data frame with Genotype, BLUP, SE, CI_lower, CI_upper
}
\description{
Compute BLUPs (genotype as random, environment as fixed)
}
