% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviromics.R
\name{compute_covariate_correlations}
\alias{compute_covariate_correlations}
\title{Compute Covariate-Phenotype Correlations}
\usage{
compute_covariate_correlations(env_data, env_means, env_col)
}
\arguments{
\item{env_data}{Environmental covariates}

\item{env_means}{Data frame from run_finlay_wilkinson (contains env_mean)}

\item{env_col}{Column name for environments}
}
\value{
Correlation matrix
}
\description{
Compute Covariate-Phenotype Correlations
}
