% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{compute_duplicate_check}
\alias{compute_duplicate_check}
\title{Check for duplicate entries in the data}
\usage{
compute_duplicate_check(df, gen_col, env_col, rep_col)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{rep_col}{Replicate column name}
}
\value{
Data frame of duplicated rows (empty if none)
}
\description{
Check for duplicate entries in the data
}
