% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviromics.R
\name{compute_env_pca}
\alias{compute_env_pca}
\title{Compute Environmental Typology (PCA)}
\usage{
compute_env_pca(env_data, env_col)
}
\arguments{
\item{env_data}{Data frame with environmental covariates}

\item{env_col}{Column name representing environments}
}
\value{
prcomp object and formatted data
}
\description{
Compute Environmental Typology (PCA)
}
