% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{compute_er_table_fast}
\alias{compute_er_table_fast}
\title{Fast Eberhart & Russell parameters from GxE means}
\usage{
compute_er_table_fast(ge)
}
\arguments{
\item{ge}{GxE matrix object from build_ge_matrices()}
}
\value{
Data frame of Eberhart-Russell parameters
}
\description{
Fast Eberhart & Russell parameters from GxE means
}
