% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{compute_shukla_table_fast}
\alias{compute_shukla_table_fast}
\title{Fast Shukla variance from Wricke and interaction SS}
\usage{
compute_shukla_table_fast(ge, wi = NULL)
}
\arguments{
\item{ge}{GxE matrix object from build_ge_matrices()}

\item{wi}{Optional precomputed Wricke ecovalence vector}
}
\value{
Data frame of Shukla variance values
}
\description{
Fast Shukla variance from Wricke and interaction SS
}
