% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{compute_wricke_table_fast}
\alias{compute_wricke_table_fast}
\title{Fast Wricke ecovalence from additive interaction matrix}
\usage{
compute_wricke_table_fast(ge)
}
\arguments{
\item{ge}{GxE matrix object from build_ge_matrices()}
}
\value{
Data frame of Wricke ecovalence values
}
\description{
Fast Wricke ecovalence from additive interaction matrix
}
