% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{descriptive_summary}
\alias{descriptive_summary}
\title{Compute descriptive statistics grouped by a factor}
\usage{
descriptive_summary(df, group_col, traits)
}
\arguments{
\item{df}{Data frame}

\item{group_col}{Column name to group by}

\item{traits}{Character vector of trait column names}
}
\value{
Tibble with Mean, SD, Min, Max, CV\%, N per group x trait
}
\description{
Compute descriptive statistics grouped by a factor
}
