% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_design}
\alias{detect_design}
\title{Detect the experimental design from the data structure}
\usage{
detect_design(df, gen_col, env_col, rep_col, block_col = NULL)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Name of genotype column}

\item{env_col}{Name of environment column}

\item{rep_col}{Name of replicate column}

\item{block_col}{Name of block column (can be NULL)}
}
\value{
List with design_type and description
}
\description{
Detect the experimental design from the data structure
}
