% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviromics.R
\name{fetch_nasa_weather}
\alias{fetch_nasa_weather}
\title{Fetch daily weather parameters from NASA POWER API}
\usage{
fetch_nasa_weather(lat, lon, start_date, end_date)
}
\arguments{
\item{lat}{Numeric latitude}

\item{lon}{Numeric longitude}

\item{start_date}{R Date object}

\item{end_date}{R Date object}
}
\value{
A data frame with daily weather variables, or NULL if query fails
}
\description{
Fetch daily weather parameters from NASA POWER API
}
