% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviromics.R
\name{parse_date_robustly}
\alias{parse_date_robustly}
\title{Parse dates robustly from various formats}
\usage{
parse_date_robustly(date_val)
}
\arguments{
\item{date_val}{Vector or single value of dates in char, factor, numeric, or date format}
}
\value{
Date vector
}
\description{
Parse dates robustly from various formats
}
