% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{pivot_ge_means}
\alias{pivot_ge_means}
\title{Pivot data to genotype × environment means for a given trait}
\usage{
pivot_ge_means(df, gen_col, env_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column}

\item{env_col}{Environment column}

\item{trait}{Trait column}
}
\value{
Wide data frame (rows = genotypes, cols = environments)
}
\description{
Pivot data to genotype × environment means for a given trait
}
