% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boxplots}
\alias{plot_boxplots}
\title{Create boxplots for trait by grouping variable}
\usage{
plot_boxplots(df, trait, group_col, color_col = NULL, show_points = FALSE)
}
\arguments{
\item{df}{Data frame}

\item{trait}{Trait column name}

\item{group_col}{Column to group by (environment or genotype)}

\item{color_col}{Column to color by (optional, defaults to group_col)}

\item{show_points}{Logical, show jittered points}
}
\value{
ggplot object
}
\description{
Create boxplots for trait by grouping variable
}
