% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_env_correlation}
\alias{plot_env_correlation}
\title{Create an environment correlation matrix plot}
\usage{
plot_env_correlation(df, gen_col, env_col, trait, method = "pearson")
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column}

\item{env_col}{Environment column}

\item{trait}{Trait column}

\item{method}{Correlation method: "pearson", "spearman", or "kendall"}
}
\value{
plotly object
}
\description{
Create an environment correlation matrix plot
}
