% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_outlier_scatter}
\alias{plot_outlier_scatter}
\title{Create an outlier detection scatter plot}
\usage{
plot_outlier_scatter(df, trait, group_col, gen_col)
}
\arguments{
\item{df}{Data frame with 'is_outlier' column}

\item{trait}{Trait column name}

\item{group_col}{Group column (usually environment)}

\item{gen_col}{Genotype column (for labeling)}
}
\value{
ggplot object
}
\description{
Create an outlier detection scatter plot
}
