% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{prepare_metan_data}
\alias{prepare_metan_data}
\title{Prepare data for stability functions}
\usage{
prepare_metan_data(df, gen_col, env_col, rep_col, trait = NULL)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{rep_col}{Replicate column name}

\item{trait}{Trait column name}
}
\value{
Data frame with cleaned factor columns
}
\description{
Prepare data for stability functions
}
