% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{read_met_file}
\alias{read_met_file}
\title{Read a MET data file (CSV or Excel)}
\usage{
read_met_file(path, sep = ",", header = TRUE)
}
\arguments{
\item{path}{File path}

\item{sep}{CSV separator (ignored for Excel)}

\item{header}{Whether first row is header}
}
\value{
Data frame
}
\description{
Read a MET data file (CSV or Excel)
}
