% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{run_all_stability}
\alias{run_all_stability}
\title{Run all stability indices}
\usage{
run_all_stability(df, gen_col, env_col, rep_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{rep_col}{Replicate column name}

\item{trait}{Trait column name}
}
\value{
List with fitted model metadata and combined stability table
}
\description{
Run all stability indices
}
