% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammi.R
\name{run_ammi}
\alias{run_ammi}
\title{Run AMMI Analysis natively (no metan dependency)}
\usage{
run_ammi(data, gen, env, rep = NULL, trait, n_axis = 2)
}
\arguments{
\item{data}{data.frame en formato largo}

\item{gen}{string: nombre de la columna de genotipo}

\item{env}{string: nombre de la columna de ambiente}

\item{rep}{string: nombre de la columna de repetición (o NULL si no hay)}

\item{trait}{string: nombre de la columna del rasgo fenotípico}

\item{n_axis}{integer: número de ejes IPCA a retener (default = 2)}
}
\value{
lista con todos los objetos necesarios para plotear y reportar
}
\description{
Run AMMI Analysis natively (no metan dependency)
}
