% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{run_ammi_metan}
\alias{run_ammi_metan}
\title{AMMI analysis using metan::performs_ammi()}
\usage{
run_ammi_metan(df, gen_col, env_col, rep_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{rep_col}{Replicate column name}

\item{trait}{Trait column name}
}
\value{
List with AMMI model, ANOVA table, and IPCA scores
}
\description{
AMMI analysis using metan::performs_ammi()
}
