% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run MERIDIAN}
\usage{
run_app(...)
}
\arguments{
\item{...}{Options passed to \code{\link[golem:with_golem_options]{golem::with_golem_options()}}.}
}
\value{
A Shiny application object.
}
\description{
Launches the MERIDIAN Shiny application.
}
