% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{run_eberhart_russell}
\alias{run_eberhart_russell}
\title{Eberhart & Russell stability regression}
\usage{
run_eberhart_russell(df, gen_col, env_col, rep_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{rep_col}{Replicate column name}

\item{trait}{Trait column name}
}
\value{
List with fitted model metadata and parameters
}
\description{
Eberhart & Russell stability regression
}
