% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviromics.R
\name{run_finlay_wilkinson}
\alias{run_finlay_wilkinson}
\title{Run Finlay-Wilkinson Reaction Norms
Computes the environmental index and fits regressions for each genotype.}
\usage{
run_finlay_wilkinson(df, gen_col, env_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{trait}{Trait column name}
}
\value{
Data frame with fitted lines and scatter data
}
\description{
Run Finlay-Wilkinson Reaction Norms
Computes the environmental index and fits regressions for each genotype.
}
