% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{run_gge}
\alias{run_gge}
\title{GGE Biplot analysis using metan::gge()}
\usage{
run_gge(df, gen_col, env_col, rep_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{rep_col}{Replicate column name}

\item{trait}{Trait column name}
}
\value{
List with GGE model and imputed-cell count
}
\description{
GGE Biplot analysis using metan::gge()
}
