% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviromics.R
\name{run_mega_envs}
\alias{run_mega_envs}
\title{Run Mega-Environment Analysis (Which-won-where)
Natively computes winning genotypes per environment and groups them.}
\usage{
run_mega_envs(df, gen_col, env_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{trait}{Trait column name}
}
\value{
List with env_strat dataframe
}
\description{
Run Mega-Environment Analysis (Which-won-where)
Natively computes winning genotypes per environment and groups them.
}
