% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_mixed_model}
\alias{run_mixed_model}
\title{Fit mixed model with lme4 and extract variance components
Model: trait ~ env + (1|gen) + (1|gen:env) + (1|env:rep)}
\usage{
run_mixed_model(df, gen_col, env_col, rep_col, trait)
}
\arguments{
\item{df}{Data frame}

\item{gen_col}{Genotype column name}

\item{env_col}{Environment column name}

\item{rep_col}{Replicate column name}

\item{trait}{Trait column name}
}
\value{
List with var_components, model, heritability
}
\description{
Fit mixed model with lme4 and extract variance components
Model: trait ~ env + (1|gen) + (1|gen:env) + (1|env:rep)
}
