% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_analysis}
\alias{safe_analysis}
\title{Execute an expression with tryCatch, showing Shiny notifications on error/warning}
\usage{
safe_analysis(expr, session)
}
\arguments{
\item{expr}{Expression to evaluate}

\item{session}{Shiny session object}
}
\value{
Result of expr, or NULL on error
}
\description{
Execute an expression with tryCatch, showing Shiny notifications on error/warning
}
