% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability_ranking}
\alias{stability_ranking}
\title{Combine stability rankings into one table}
\usage{
stability_ranking(stats_table)
}
\arguments{
\item{stats_table}{Stability statistics data frame}
}
\value{
Rounded stability ranking data frame
}
\description{
Combine stability rankings into one table
}
