% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_upload.R
\name{validate_environmental_data}
\alias{validate_environmental_data}
\title{Validate environmental covariates dataset}
\usage{
validate_environmental_data(env_df)
}
\arguments{
\item{env_df}{Data frame}
}
\value{
List of errors and warnings
}
\description{
Validate environmental covariates dataset
}
