% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviromics.R
\name{calculate_enviromic_indices}
\alias{calculate_enviromic_indices}
\title{Calculate summary agroclimatic indices from daily weather series}
\usage{
calculate_enviromic_indices(daily_weather)
}
\arguments{
\item{daily_weather}{Data frame containing daily variables}
}
\value{
A single-row data frame with computed indices
}
\description{
Calculate summary agroclimatic indices from daily weather series
}
